// src/app/shared/utils/enums/momentEnum.ts
export enum MomentEnum {
    LANGUAGE = 'pt-br',
    HORA = 'LT',
    HORA_SEGUNDOS = 'LTS',
    DATA = 'L',
    DATA_EXTENSO = 'LL',
    DATA_ABREV = 'll',
    DATA_COMPLETA = 'LLLL',
    YEARS = 'years',
    MONTHS = 'months',
    WEEKS = 'weeks',
    DAYS = 'days',
    HOURS = 'hours',
    MINUTES = 'minutes',
    SECONDS = 'seconds',
    MILLISECONDS = 'milliseconds',
}

// Documentation https://momentjs.com/